added multidirectional move function to chassis - not tested yet
